code update

+++ admin/src/main/java/com/nycloud/admin/dto/PermissionResourceDto.java
@@ -15,6 +15,6 @@ public class PermissionResourceDto extends RequestDto{
 
     private Integer permissionId;
 
-    private Integer [] resourceIds;
+        private Long [] resourceIds;
 
 }
